+++ app/app/(app)/(drawer)/(ride)/_layout.tsx
@@ -1,0 +1,7 @@
+import { Stack } from "expo-router"
+
+export default function Layout() {
+  return (
+    <Stack screenOptions={{ headerShown: false, headerTransparent: true, headerTitle: '' }} />
+  );
+}
